+++ apps/meteor/tests/e2e/omnichannel/omnichannel-contact-center-chats.spec.ts
@@ -5,6 +5,7 @@ import { IS_EE } from '../config/constants';
 import { Users } from '../fixtures/userStates';
 import { HomeOmnichannel } from '../page-objects';
 import { OmnichannelContactCenterChats } from '../page-objects/omnichannel';
+import { setSettingValueById } from '../utils';
 import { createAgent, makeAgentAvailable } from '../utils/omnichannel/agents';
 import { addAgentToDepartment, createDepartment } from '../utils/omnichannel/departments';
 import { createConversation, updateRoom } from '../utils/omnichannel/rooms';
@@ -108,6 +109,7 @@ test.describe('OC - Contact Center Chats [Auto Selection]', async () => {
 				tags: [tagB.data.name],
 			}),
 		]);
+		await setSettingValueById(api, 'Omnichannel_enable_department_removal', true);
 	});
 
 	test.beforeEach(async ({ page }: { page: Page }) => {
@@ -132,6 +134,7 @@ test.describe('OC - Contact Center Chats [Auto Selection]', async () => {
 			api.post('/settings/Livechat_allow_manual_on_hold', { value: false }),
 			api.post('/settings/Livechat_allow_manual_on_hold_upon_agent_engagement_only', { value: true }),
 		]);
+		await setSettingValueById(api, 'Omnichannel_enable_department_removal', false);
 	});
 
 	// Change conversation A to on hold and close conversation B